+++ examples/show-index.c
@@ -17,7 +17,7 @@
 int lg2_show_index(git_repository *repo, int argc, char** argv)
 {
 	git_index *index;
-	unsigned int i, ecount;
+	size_t i, ecount;
 	char *dir = ".";
 	size_t dirlen;
 	char out[GIT_OID_HEXSZ+1];
